Refacter vmdetails and added get_vm_details
