+++ drivers/aws_shell/src/driver.py
@@ -63,3 +63,6 @@ class AWSShellDriver(ResourceDriverInterface):
 
     def SetAppSecurityGroups(self, context, request):
         return self.aws_shell.set_app_security_groups(context, request)
+
+    def GetVmDetails(self, context, ports):
+        return self.aws_shell.get_vm_details(context)
